requirements.txt
<<<<<<< HEAD
Flask==3.0.0
Flask-Session==0.6.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
=======
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.0
supabase==2.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
>>>>>>> 20dac40646ecb456f1bdc39aa36c2952699ee397
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
<<<<<<< HEAD
blinker==1.7.0
bcrypt==4.1.2
WTForms==3.1.1
=======
blinker==1.6.2
bcrypt==4.1.2
>>>>>>> 20dac40646ecb456f1bdc39aa36c2952699ee397
